# front/src/_generated/typescript-fetch/models/FilePaths.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FilePaths
 */
export interface FilePaths {
    /**
     * 
     * @type {string}
     * @memberof FilePaths
     */
    u2: string;
    /**
     * 
     * @type {string}
     * @memberof FilePaths
     */
    v2: string;
    /**
     * 
     * @type {string}
     * @memberof FilePaths
     */
    w2: string;
    /**
     * 
     * @type {string}
     * @memberof FilePaths
     */
    theta: string;
}

/**
 * Check if a given object implements the FilePaths interface.
 */
export function instanceOfFilePaths(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "u2" in value;
    isInstance = isInstance && "v2" in value;
    isInstance = isInstance && "w2" in value;
    isInstance = isInstance && "theta" in value;

    return isInstance;
}

export function FilePathsFromJSON(json: any): FilePaths {
    return FilePathsFromJSONTyped(json, false);
}

export function FilePathsFromJSONTyped(json: any, ignoreDiscriminator: boolean): FilePaths {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'u2': json['u2'],
        'v2': json['v2'],
        'w2': json['w2'],
        'theta': json['theta'],
    };
}

export function FilePathsToJSON(value?: FilePaths | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'u2': value.u2,
        'v2': value.v2,
        'w2': value.w2,
        'theta': value.theta,
    };
}