# openapi.json
{
 "openapi": "3.0.3",
 "info": {
  "title": "FastAPI",
  "version": "0.1.0"
 },
 "paths": {
  "/upload": {
   "post": {
    "summary": "Upload File",
    "operationId": "upload_file_upload_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_file_upload_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": {
          "type": "string"
         },
         "type": "object",
         "title": "Response Upload File Upload Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/__ws_placeholder": {
   "post": {
    "summary": "Ws Placeholder",
    "operationId": "ws_placeholder___ws_placeholder_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/WsData"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  }
 },
 "components": {
  "schemas": {
   "Body_upload_file_upload_post": {
    "properties": {
     "u2": {
      "type": "string",
      "format": "binary",
      "title": "U2"
     },
     "v2": {
      "type": "string",
      "format": "binary",
      "title": "V2"
     },
     "w2": {
      "type": "string",
      "format": "binary",
      "title": "W2"
     },
     "theta": {
      "type": "string",
      "format": "binary",
      "title": "Theta"
     }
    },
    "type": "object",
    "required": [
     "u2",
     "v2",
     "w2",
     "theta"
    ],
    "title": "Body_upload_file_upload_post"
   },
   "FilePaths": {
    "properties": {
     "u2": {
      "type": "string",
      "title": "U2"
     },
     "v2": {
      "type": "string",
      "title": "V2"
     },
     "w2": {
      "type": "string",
      "title": "W2"
     },
     "theta": {
      "type": "string",
      "title": "Theta"
     }
    },
    "type": "object",
    "required": [
     "u2",
     "v2",
     "w2",
     "theta"
    ],
    "title": "FilePaths"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
   "WsData": {
    "properties": {
     "action": {
      "type": "string",
      "title": "Action"
     },
     "file_paths": {
      "$ref": "#/components/schemas/FilePaths"
     },
     "fs": {
      "type": "number",
      "title": "Fs"
     },
     "fw1": {
      "type": "number",
      "title": "Fw1"
     },
     "fw2": {
      "type": "number",
      "title": "Fw2"
     }
    },
    "type": "object",
    "required": [
     "action",
     "file_paths",
     "fs",
     "fw1",
     "fw2"
    ],
    "title": "WsData"
   }
  }
 }
}